added_fix
refined code for videoplayer
added check for empty fields in the form

+++ src/js/03-feedback.js
@@ -22,10 +22,16 @@ function onFormInput(evt) {
 
 function onFormSubmit(evt) {
 
+    if (!email.value || !message.value){
+        alert('check that all fields are filled');
+        return;
+    }
+
     console.log('localStorage:', JSON.parse(localStorage.getItem("feedback-form-state")));
     evt.preventDefault();
     evt.currentTarget.reset();
     localStorage.clear();
+
 }
 
 function updateOutput() {
